requirements.txt
autopep8==2.0.2
kafka-python==2.0.2
msgpack==1.0.5
pycodestyle==2.10.0
